gateway/ai_service/requirements.txt

-e .
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
blis==0.7.11
catalogue==2.0.10
celery==5.5.3
certifi==2025.4.26
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
daphne==4.0.0
datasets==2.14.4
dill==0.3.7
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fasttext-wheel==0.9.2
ffmpeg-python==0.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
gunicorn==20.1.0
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.15
murmurhash==1.0.13
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
packaging==25.0
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.2.1
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pybind11==2.13.6
pydantic==1.10.22
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
spacy==3.8.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.13.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.29.2
typer==0.7.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
whisper==1.1.10
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
